+++ src/utils.ts
@@ -90,14 +90,50 @@ export function wrapCrashReporting(f) {
 }
 
 export function resolvePath(p: string, normalize: boolean = true) {
-    p = parseEnvVariables(p)
+    p = parseVSCodeVariables(p)
     p = p.replace(/^~/, os.homedir())
     p = normalize ? path.normalize(p) : p
     return p
 }
 
-function parseEnvVariables(p: string) {
-    return p.replace(/\${env:(.*?)}/g, (_, variable) => {
+/**
+ * Parse a subset of VSCode 'variables' in `p`, and return a string with the replacements.
+ *
+ * Specifically, we support:
+ *  - ${userHome}
+ *  - ${workspaceFolder}
+ *  - ${workspaceFolder:<FOLDER_NAME>}  (For a multi-root project, use the first folder)
+ *  - ${pathSeparator}
+ *  - ${env:<ENVIRONMENT_VARIABLE>}
+ *
+ * See https://code.visualstudio.com/docs/editor/variables-reference for definitions of the
+ * above.
+ *
+ * TODO: this replicates functionality present in core VSCode! The implementation of this
+ *  function be replaced once this issue is resolved:
+ *      https://github.com/microsoft/vscode/issues/46471
+ */
+function parseVSCodeVariables(p: string) {
+    p = p.replace(/\${userHome}/g, os.homedir())
+
+    const workspace_paths = vscode.workspace.workspaceFolders.map((folder) => {
+        return folder.uri.fsPath
+    })
+    p = p.replace(/\${workspaceFolder}/g, (_) => {
+        // In the case of a multi-root workspace, we return the first one.
+        return workspace_paths.length >= 1 ? workspace_paths[0] : null
+    })
+    p = p.replace(/\${workspaceFolder:(.*?)}/g, (_, desired_basename) => {
+        const filtered_paths = workspace_paths.filter((workspace_path) => {
+            return desired_basename === path.basename(workspace_path)
+        })
+        // If we have zero or more than one matches, we cannot proceed.
+        return filtered_paths.length === 1 ? filtered_paths[0] : null
+    })
+
+    p = p.replace(/\${pathSeparator}/g, path.sep)
+    p = p.replace(/\${env:(.*?)}/g, (_, variable) => {
         return process.env[variable] || ''
     })
+    return p
 }