+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to the Julia extension will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+* Support additional VSCode variables in `environmentPath` ([#3477](https://github.com/julia-vscode/julia-vscode/pull/3477))
+
 ### Fixed
 * Table headers no longer show `null` if no column label existss ([#3486](https://github.com/julia-vscode/julia-vscode/pull/3486))